Reorganization done to a satisfactory level

+++ tasks/forwardData.js
@@ -1,11 +1,13 @@
 function createFollowDataTask(execlib){
   'use strict';
   var lib = execlib.lib,
-      q = lib.q,
-      execSuite = execlib.execSuite,
-      MultiDestroyableTask = execSuite.MultiDestroyableTask,
-      dataSuite = execlib.dataSuite,
-      DataDecoder = dataSuite.DataDecoder;
+    q = lib.q,
+    qlib = lib.qlib,
+    execSuite = execlib.execSuite,
+    taskRegistry = execSuite.taskRegistry,
+    MultiDestroyableTask = execSuite.MultiDestroyableTask,
+    dataSuite = execlib.dataSuite,
+    DataDecoder = dataSuite.DataDecoder;
 
   function ChildSinkStorage(sink){
     this.sink = sink;
@@ -26,21 +28,42 @@ function createFollowDataTask(execlib){
 
   function FollowDataTask(prophash){
     MultiDestroyableTask.call(this,prophash,['sink','childsink']);
-    this.storage = new ChildSinkStorage(prophash.childsink);
     this.sink = prophash.sink;
+    this.storage = new ChildSinkStorage(prophash.childsink);
+    this.decoder = null;
   }
   lib.inherit(FollowDataTask,MultiDestroyableTask);
   FollowDataTask.prototype.__cleanUp = function(){
-    this.storage.destroy();
-    this.storage = null;
+    /*
     if(this.sink.destroyed){ //it's still alive
       this.sink.consumeChannel('d',lib.dummyFunc);
     }
+    */
+    if (this.sink && this.decoder && this.decoder.queryID) {
+      this.sink.sessionCall('closeQuery', this.decoder.queryID);
+    }
+    if (this.decoder) {
+      this.decoder.destroy();
+    }
+    this.decoder = null;
+    if (this.storage) {
+      this.storage.destroy();
+    }
+    this.storage = null;
     this.sink = null;
     MultiDestroyableTask.prototype.__cleanUp.call(this);
   };
   FollowDataTask.prototype.go = function(){
-    this.sink.consumeChannel('d',new DataDecoder(this.storage));
+    if (this.decoder) {
+      return;
+    }
+    this.decoder = new DataDecoder(this.storage); 
+    this.sink.sessionCall('query', {continuous:true}).then(
+      this.destroy.bind(this),
+      this.destroy.bind(this),
+      this.decoder.onStream.bind(this.decoder)
+    );
+    //this.sink.consumeChannel('d',new DataDecoder(this.storage));
   };
   FollowDataTask.prototype.compulsoryConstructionProperties = ['sink','childsink'];
